Needed to expanded the attributes on the grammar class.
Realized that I wanted the strings returned from the lexer to have
some additional functions (like str.isnumeric()) to help test whether
they match certain parts of the grammar, so implemented the token class.

Streamlined the lexer a bit to handle comments differently, and also
to deal with reserved characters that start a numeric.