Update pull_request_template.md
Changed these elements from "Task List" to just plain bulleted lists and improved instructions.  When they are left as "Task Lists" then they can be accidentally checked via misclick, and since not all of these will be checked, it constantly shows that some of the 'tasks' on this PR are not done in references to the PR, which is misleading.